+++ src/create-layer.ts
@@ -25,7 +25,7 @@ export function createLayer(
       toFragmentFromInstance(useConfig),
       Content ? { content: { component: Content } } : undefined,
     )
-    const { instance } = createLayerInstance({ create, adapter, use })
+    const instance = createLayerInstance({ create, adapter, use })
     instance.bindHost()
     return instance
   }